Reduced dependency on subscription and paho-mqtt when not necessary
// includes all the necessary dependencies
import AWSAppSyncClient from "aws-appsync";
import { authIAM } from "aws-appsync/lib/link/auth/aws";
import { ComplexObjectLink } from "aws-appsync/lib/link/complex";
import { subscription } from "aws-appsync/lib/link/subscription";
const client = new AWSAppSyncClient({
url: 'some url',
region: 'some region',
authType: authIAM(credentials),
complexObjects: new ComplexObjectLink(complexObjectsCredentials),
subscription: subscription,
})
