+++ packages/aws-appsync/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aws-appsync",
-  "version": "1.7.1-b",
+  "version": "1.7.1-b02",
   "main": "lib/index.js",
   "license": "SEE LICENSE IN LICENSE",
   "description": "AWS Mobile AppSync SDK for JavaScript",
