+++ packages/aws-appsync/src/deltaSync.ts
@@ -22,7 +22,7 @@ import { SKIP_RETRY_KEY } from "./link/retry-link";
 import { DocumentNode, print, OperationDefinitionNode, FieldNode, ExecutionResult } from "graphql";
 import { getOpTypeFromOperationName, CacheOperationTypes, getUpdater, QueryWithVariables } from "./helpers/offline";
 import { boundSaveSnapshot, replaceUsingMap, EnqueuedMutationEffect, offlineEffectConfig as mutationsConfig } from "./link/offline-link";
-import { CONTROL_EVENTS_KEY } from "./link/subscription-handshake-link";
+import { CONTROL_EVENTS_KEY } from "./utils";
 
 const logger = rootLogger.extend('deltasync');
