+++ packages/aws-appsync/src/link/auth/auth-link.ts
@@ -37,7 +37,6 @@ export class AuthLink extends ApolloLink {
                 ...extraHeaders
             },
         });
-
         return forward(operation);
     };
