+++ packages/aws-appsync/src/utils/index.ts
@@ -10,6 +10,8 @@ import { DocumentNode, OperationDefinitionNode, FieldNode } from "graphql";
 import { resultKeyNameFromField } from "apollo-utilities";
 import { Observable } from "apollo-link";
 
+export const CONTROL_EVENTS_KEY = '@@controlEvents';
+
 const crypto = require('aws-sdk/lib/browserCryptoLib');
 
 export const passthroughLink = (op, forward) => (forward ? forward(op) : Observable.of());
